"Refactored project structure and updated documentation"
Relocated solution*.adoc files inside the src/site/asciidoc directory for enhanced project structure readability. The change was needed to conform to a standardized project layout. Updated references to these files accordingly.

Added information about the developer and the license under which the project is published to the pom.xml. Also significantly enhanced the site's publishing configuration by adding more attributes, ensuring correct character encoding, and adding asciidoctor-maven-plugin as a dependency.

Added files and writes up for the sixth task: Producer-Consumer multi-threading. This includes a task file, which provides a write-up of the task and its solution and the source code of the Classic and Concurrency solutions.

Also committed are two Java files offering implementation solution to the problem statement in producer-consumer multi-threading. Corrections made to Maven dependencies for compatibility in the pom.xml file.